Rename the `modules` directory to `kodos`
