+++ kodos.pro
@@ -1,31 +1,31 @@
 SOURCES = kodos
-	modules/about.py
-	modules/debug.py
-	modules/help.py
-	modules/__init__.py
-	modules/parseRegexLib.py
-	modules/prefs.py
-	modules/recent_files.py
-	modules/reference.py
-	modules/regexLibrary.py
-	modules/reportBug.py
-	modules/status_bar.py
-	modules/tooltip.py
-	modules/urlDialog.py
-	modules/util.py
-	modules/version.py
+	kodos/about.py
+	kodos/debug.py
+	kodos/help.py
+	kodos/__init__.py
+	kodos/parseRegexLib.py
+	kodos/prefs.py
+	kodos/recent_files.py
+	kodos/reference.py
+	kodos/regexLibrary.py
+	kodos/reportBug.py
+	kodos/status_bar.py
+	kodos/tooltip.py
+	kodos/urlDialog.py
+	kodos/util.py
+	kodos/version.py
 
 TRANSLATIONS =	translations/kodos_en.ts \
 		translations/kodos_pl.ts \
 		translations/kodos_sv.ts
 
-FORMS =	modules/aboutBA.ui \
-	modules/helpBA.ui \
-	modules/kodosBA.ui \
-	modules/newUserDialogBA.ui
-	modules/prefsBA.ui \
-	modules/referenceBA.ui \
-	modules/regexLibraryBA.ui \
-	modules/reportBugBA.ui \
-	modules/resultsBA.ui \
-	modules/urlDialogBA.ui
+FORMS =	kodos/aboutBA.ui \
+	kodos/helpBA.ui \
+	kodos/kodosBA.ui \
+	kodos/newUserDialogBA.ui
+	kodos/prefsBA.ui \
+	kodos/referenceBA.ui \
+	kodos/regexLibraryBA.ui \
+	kodos/reportBugBA.ui \
+	kodos/resultsBA.ui \
+	kodos/urlDialogBA.ui
